:recycle: allow get_key and get_quote for linux only$
